+++ src/routes/users/useUsersStore.tsx
@@ -5,45 +5,71 @@ import {
     createUser,
     deleteUser,
     listUsers,
+    listUsersAwaitingApproval,
     retrieveUser,
     updateUser,
+    deleteUserAwaitingApproval,
+    approveUserAwaitingApproval,
 } from '@/api/UsersApi';
 import { handleUnknownError, withAbortSignal } from '@/api/request';
 import { APIListArgsT, AbortFunctionT } from '@/types/CommonTypes';
-import { UserT, UpdateUserPayloadT, UserPayloadT } from '@/types/UsersTypes';
+import {
+    UserT,
+    UpdateUserPayloadT,
+    UserPayloadT,
+    UserApprovalPayloadT,
+} from '@/types/UsersTypes';
 
 type UsersStateT = {
     users: UserT[];
+    usersAwaitingApproval: UserT[];
     usersCount: number;
+    usersAwaitingApprovalCount: number;
     user: UserT | null;
+    userAwaitingApproval: UserT | null;
 
     fetchingUsers: boolean;
+    fetchingUsersAwaitingApproval: boolean;
     fetchingUser: boolean;
     creatingUser: boolean;
     updatingUser: boolean;
+    approvingUserAwaitingApproval: boolean;
     deletingUser: boolean;
+    deletingUserAwaitingApproval: boolean;
 
     fetchUsers: (params: APIListArgsT) => AbortFunctionT;
+    fetchUsersAwaitingApproval: () => AbortFunctionT;
     fetchUser: (key: string) => Promise<UserT | null>;
     createUser: (payload: UserPayloadT) => Promise<string | null>;
     updateUser: (
         key: string,
         payload: UpdateUserPayloadT
     ) => Promise<string | null>;
+    approveUserAwaitingApproval: (
+        key: string,
+        payload: UserApprovalPayloadT
+    ) => Promise<string | null>;
     deleteUser: (key: string) => Promise<string | null>;
+    deleteUserAwaitingApproval: (key: string) => Promise<string | null>;
 };
 
 let fetchUserController: AbortController | null;
 
 const useUsersStore = create<UsersStateT>((set, get) => ({
     users: [],
+    usersAwaitingApproval: [],
     usersCount: 0,
+    usersAwaitingApprovalCount: 0,
     user: null,
+    userAwaitingApproval: null,
     fetchingUsers: true,
+    fetchingUsersAwaitingApproval: true,
     fetchingUser: true,
     creatingUser: false,
     updatingUser: false,
+    approvingUserAwaitingApproval: false,
     deletingUser: false,
+    deletingUserAwaitingApproval: false,
     fetchUsers: withAbortSignal(async (signal, params) => {
         set({ fetchingUsers: true });
         try {
@@ -61,6 +87,24 @@ const useUsersStore = create<UsersStateT>((set, get) => ({
             } else {
                 console.warn(error);
                 set({ fetchingUsers: false });
+            }
+        }
+    }),
+    fetchUsersAwaitingApproval: withAbortSignal(async (signal) => {
+        set({ fetchingUsersAwaitingApproval: true });
+        try {
+            const response = await listUsersAwaitingApproval({ signal });
+            set({
+                fetchingUsersAwaitingApproval: false,
+                usersAwaitingApproval: response.data.results,
+                usersAwaitingApprovalCount: response.data.count,
+            });
+        } catch (error) {
+            if (axios.isCancel(error)) {
+                // do nothing, call has been canceled voluntarily
+            } else {
+                console.warn(error);
+                set({ fetchingUsersAwaitingApproval: false });
             }
         }
     }),
@@ -125,6 +169,27 @@ const useUsersStore = create<UsersStateT>((set, get) => ({
             return handleUnknownError(error);
         }
     },
+    approveUserAwaitingApproval: async (
+        key: string,
+        payload: UserApprovalPayloadT
+    ) => {
+        set({ approvingUserAwaitingApproval: true });
+        try {
+            const response = await approveUserAwaitingApproval(key, payload);
+            set({
+                usersAwaitingApprovalCount:
+                    get().usersAwaitingApprovalCount - 1,
+                approvingUserAwaitingApproval: false,
+                usersAwaitingApproval: get().usersAwaitingApproval.map((user) =>
+                    user.username === key ? response.data : user
+                ),
+            });
+            return null;
+        } catch (error) {
+            set({ approvingUserAwaitingApproval: false });
+            return handleUnknownError(error);
+        }
+    },
     deleteUser: async (key: string) => {
         set({ deletingUser: true });
         try {
@@ -133,6 +198,25 @@ const useUsersStore = create<UsersStateT>((set, get) => ({
             return null;
         } catch (error) {
             set({ deletingUser: false });
+            return handleUnknownError(error);
+        }
+    },
+    deleteUserAwaitingApproval: async (key: string) => {
+        set({ deletingUserAwaitingApproval: true });
+        try {
+            await deleteUserAwaitingApproval(key);
+            set({
+                deletingUserAwaitingApproval: false,
+                usersAwaitingApprovalCount:
+                    get().usersAwaitingApprovalCount - 1,
+                usersAwaitingApproval: get().usersAwaitingApproval.filter(
+                    (userAwaitingApproval) =>
+                        userAwaitingApproval.username !==
+                        get().userAwaitingApproval?.username
+                ),
+            });
+            return null;
+        } catch (error) {
             return handleUnknownError(error);
         }
     },